Test script, larger cache.
I've bumped the in-SRAM cache to 256 bytes. When I solder up an actual
SRAM, I'll have a different infrastructure. For now, this is using the
SRAM on the ATmega328p.

The test script is running on a CircuitPython Express.

I should have a makefile for copying it over to the CPX.